sample/pyproject.toml
[project]
name = "research-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.25",
    "langchain-mcp-adapters>=0.1.7",
    "langchain-ollama>=0.3.3",
    "langchain-openai>=0.3.23",
    "langgraph>=0.4.8",
    "mcp>=1.9.4",
    "python-dotenv>=1.1.0",
]
